pyproject.toml
[project]
name = "awesome-inc"
version = "0.4.1"
description = "Add your description here"
authors = [
    { name = "nemishmehta", email = "23.nemishmehta@gmail.com" }
]
dependencies = [
    "fastapi>=0.111.1",
    "sqlalchemy>=2.0.31",
    "pydantic>=2.8.2",
    "psycopg2>=2.9.9",
    "httpx>=0.27.0",
    "dbt-core>=1.8.4",
    "dbt-postgres>=1.8.2",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "re-data>=0.11.0",
    "pytest-mock>=3.14.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.1",
    "mypy>=1.11.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.1",
    "ruff>=0.5.4",
]

[tool.ruff]
line-length = 79

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/awesome_inc"]
